[common] Align bitmap writes with external sort
